fix: make chunked append configurable and move box helpers to mp4-tools
Move fMP4 box-boundary splitting functions (`readMp4BoxSize`,
`splitAtBoxBoundaries`, `splitAppendData`) from buffer-controller to
mp4-tools where they belong.

Replace hardcoded `MAX_APPEND_SIZE` constant with a new
`config.maxAppendSize` option (default: `Infinity`) so the feature is
opt-in and the chunk size is user-configurable. Splitting only activates
when `maxAppendSize` is finite.
