+++ api-extractor/report/hls.js.api.md
@@ -786,6 +786,7 @@ export type BufferControllerConfig = {
     loopBackBufferFlush?: boolean;
     liveDurationInfinity: boolean;
     liveBackBufferLength: number | null;
+    maxAppendSize: number;
 };
 
 // Warning: (ae-missing-release-tag) "BufferCreatedData" is part of the package's API, but it is missing a release tag (@alpha, @beta, @public, or @internal)
