+++ src/config.ts
@@ -82,6 +82,7 @@ export type BufferControllerConfig = {
    * @deprecated use backBufferLength
    */
   liveBackBufferLength: number | null;
+  maxAppendSize: number;
 };
 
 export type CapLevelControllerConfig = {
@@ -456,6 +457,7 @@ export const hlsDefaultConfig: HlsConfig = {
   backBufferLength: Infinity, // used by buffer-controller
   frontBufferFlushThreshold: Infinity,
   loopBackBufferFlush: undefined, // used by buffer-controller
+  maxAppendSize: Infinity, // used by buffer-controller
   startOnSegmentBoundary: false, // used by stream-controller
   nextAudioTrackBufferFlushForwardOffset: 0.25, // used by stream-controller
   maxBufferSize: 60 * 1000 * 1000, // used by stream-controller
